FinanceTracker/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "firebase-admin>=6.7.0",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "pyrebase4>=4.8.0",
    "streamlit>=1.43.2",
]
